Broke the modules into apiCall and render, and css changes
