+++ src/style.css
@@ -40,6 +40,7 @@ hr {
   flex-wrap: wrap;
   justify-content: space-around;
   padding: 0;
+  margin-top: 20px;
 }
 
 .card {
